+++ CHANGELOG.md
@@ -7,9 +7,14 @@ Versioning].
 
 ## [Unreleased]
 
+- Disable the action from a Pull Request comment. ([#106])
+
+## [0.2.2] - 2020-03-02
+
 - Configure the SVGO options file. ([#99])
 - Fix `required` value for "configuration-path" input. ([#100])
 - Fix interpretation of "dry-run" input in config file. ([#103])
+- Fix documentation for configuration in config file. ([#108])
 
 ## [0.2.1] - 2020-03-02
 
@@ -58,3 +63,5 @@ Versioning].
 [#99]: https://github.com/ericcornelissen/svgo-action/pull/99
 [#100]: https://github.com/ericcornelissen/svgo-action/pull/100
 [#103]: https://github.com/ericcornelissen/svgo-action/pull/103
+[#106]: https://github.com/ericcornelissen/svgo-action/pull/106
+[#108]: https://github.com/ericcornelissen/svgo-action/pull/108